Separates components
